Finish barcode generation. Capped at 50 for performance.
Turns out client-side is probably the wrong place for this. Although
it all technically works (which I wasn't expecting to be possible),
it's too slow to generate 1,600 barcodes on pageload. So this needs
to be pushed to the server. Shouldn't be any problems with that -
PDFKit generates a node stream anyway, and my fork of JSBarcode
doesn't really use any browser stuff (apart from some jQuery which
should be stripped away anyway imo)